requirements.txt
langchain
faiss-cpu
PyMuPDF
sentence-transformers
openai
numpy
streamlit